Add legend; refactor into components
